float -> double

+++ main.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 const int N = 16;
-Imagine::Matrix<float> victoryMatrix(N,N);
+Imagine::Matrix<double> victoryMatrix(N,N);
 
 int main()
 {
